- made output more compliant with old version - added PerSCorer.h and BleuScorer.h - stored feature names - fixed bug about output of best Point
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder@1796 1f5c12ca-751b-0410-a591-d2e778427230

+++ trunk/mert/BleuScorer.cpp
@@ -1,7 +1,6 @@
-#include "Scorer.h"
+#include "BleuScorer.h"
 
 const int BleuScorer::LENGTH = 4;
-
 
 
 /**
@@ -90,9 +89,9 @@ void BleuScorer::setReferenceFiles(const vector<string>& referenceFiles) {
 }
 
 
-void BleuScorer::prepareStats(unsigned int sid, const string& text, ScoreStats& entry) {
-        //cerr << text << endl;
-        //dump_counts(*_refcounts[sid]);
+void BleuScorer::prepareStats(size_t sid, const string& text, ScoreStats& entry) {
+    //cerr << text << endl;
+   //dump_counts(*_refcounts[sid]);
 		if (sid >= _refcounts.size()) {
 			stringstream msg;
 			msg << "Sentence id (" << sid << ") not found in reference set";
